requirements.txt
python-magic==0.4.27
python-magic-bin==0.4.14
Pillow==9.4.0